Create a new file for SpaceButtonStyle structure
