+++ solutions/devsprint-eduardo-bocato-1/Buttons/Styles/SpaceButtonStyle.swift
@@ -1,0 +1,18 @@
+//
+//  SpaceButtonStyle.swift
+//  SpaceApp
+//
+//  Created by Razee Hussein-Jamal on 10/03/2022.
+//
+
+import SwiftUI
+
+struct SpaceButtonStyle: ButtonStyle {
+    func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .padding()
+            .background(.blue)
+            .foregroundColor(.white)
+            .cornerRadius(15)
+    }
+}
